Extracted captures into its own file
